Uploaded Secret Ship Messages
This is a executable which copies the LSD from one png, to the other. This can be used in Cosmoteer to get a png of a ship and another one, and then to secretly make one ship hold the LSDs of the pixels from the other one. When the one which secretly holds the other one is plopped into Cosmoteer, Cosmoteer will actually make the ship which is disguised in the one secretly holding it. This can be usefull for secret messages. Report if it does not work!