Source code for Secret Ship Messages
This is the source code of an little project which I made with my friend Kokice5! Any issues can be reported. The code is supposed to copy the LSDs of the pixels from one image to the other, and make a new png with the new png with the applied effects.